more cleanup of filename and description text
"GitHub for Windows" has been very insistent on deleting
"FAQ/What-is-the-VRT's-Role.md" and i think i finally have made it
happy. originally it was not even pulling the file to the local
repository which is why i even messed with the name at all. now that i
got the file into the local repository, i've renamed it as well as
adjusted the two READMEs as necessary. that should finally make it
happy. i believe that the problem is/was the "?" in the original
filename as well as the "'" (apostrophe).